update part A to correct delete
In the case when the entry to delete has two children, after checking
that they are not null, we also check if they are not external
Also... there might be a case when we try to delete the largest item and
when this lies at the root.
In this case, the successor will be null, so we need to make the new
root the left child.